minimizing the code surface for easier maintainance
