+++ nemo_rl/distributed/virtual_cluster.py
@@ -1095,5 +1095,4 @@ class RayVirtualCluster:
         the cluster is lost due to leaving a function scope. It's always recommended that the
         user calls shutdown().
         """
-        if not sys.is_finalizing():
-            self.shutdown()
+        self.shutdown()
