+++ nemo_rl/models/generation/vllm/vllm_worker_async.py
@@ -218,17 +218,6 @@ class VllmAsyncGenerationWorkerImpl(BaseVllmGenerationWorker):
         self.llm = None
         self.vllm_device_ids = None
 
-    def _refit_collective_rpc(
-        self,
-        method: str,
-        args: tuple[Any, ...],
-    ) -> Any:
-        if self._refit_async_loop is None:
-            raise RuntimeError("The async vLLM refit server loop is not initialized.")
-        return asyncio.run_coroutine_threadsafe(
-            self.llm.collective_rpc(method, args=args), self._refit_async_loop
-        ).result()
-
     def _return_routed_experts_enabled(self) -> bool:
         engine_args = getattr(self, "llm_async_engine_args", None)
         if bool(getattr(engine_args, "enable_return_routed_experts", False)):
@@ -440,9 +429,9 @@ class VllmAsyncGenerationWorkerImpl(BaseVllmGenerationWorker):
             await self.llm.collective_rpc(
                 "load_mtp_weights_from_disk", args=(self.model_name,)
             )
-        if self.cfg.get("refit_transport") is not None:
+        if self._sparse_refit_receiver is not None:
             hostnames = await self.llm.collective_rpc("report_node_hostname", args=())
-            self._refit_workers_share_node = len(set(hostnames)) == 1
+            self._sparse_refit_receiver.set_worker_hostnames(hostnames)
 
     async def get_reserved_url(self) -> Optional[str]:
         """Return the URL from the reserved socket, available before model loading."""
@@ -930,8 +919,8 @@ class VllmAsyncGenerationWorkerImpl(BaseVllmGenerationWorker):
         app = FastAPI()
 
         app = self._setup_vllm_openai_api_server(app)
-        if self.cfg.get("refit_transport") is not None:
-            self._setup_vllm_refit_api_server(app)
+        if self._sparse_refit_receiver is not None:
+            self._sparse_refit_receiver.setup_api_server(app)
 
         ########################################
         # Server spinup
@@ -1556,9 +1545,8 @@ class VllmAsyncGenerationWorkerImpl(BaseVllmGenerationWorker):
     async def shutdown(self) -> bool:
         """Clean up vLLM resources."""
         try:
-            self.stop_zmq_sparse_refit_relay()
-            await asyncio.to_thread(self._flush_queued_sparse_payloads)
-            self._refit_apply_executor.shutdown(wait=True)
+            if self._sparse_refit_receiver is not None:
+                await asyncio.to_thread(self._sparse_refit_receiver.shutdown)
 
             if self.llm is not None:
                 # Clean up extension resources (e.g., ZMQ sockets)
