+++ nemo_rl/models/policy/lm_policy.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
-import sys
 import warnings
 from collections import defaultdict
 from contextlib import nullcontext
@@ -941,13 +940,6 @@ class Policy(ColocatablePolicyInterface, GenerationInterface):
         # Only get the first worker's info since all workers will have the same result
         return results[0]
 
-    def init_remote_sparse_delta_baseline(self, transport: str) -> list[ray.ObjectRef]:
-        """Initialize source-side sparse-delta baselines for remote refit."""
-        return self._run_remote_sparse_refit_workers(
-            "init_remote_sparse_delta_baseline",
-            transport=transport,
-        )
-
     def finish_inference(self) -> None:
         """Offload policy model to CPU after inference."""
         futures = self.worker_group.run_all_workers_single_data("finish_inference")
@@ -1071,45 +1063,6 @@ class Policy(ColocatablePolicyInterface, GenerationInterface):
             )
         )
 
-    def stream_remote_sparse_weights(
-        self,
-        transport: str,
-        targets: list[str],
-        *,
-        transfer_id: str,
-        api_key_env_var: Optional[str],
-        timeout_s: float,
-    ) -> list[ray.ObjectRef]:
-        """Stream sparse deltas through the selected remote value plane."""
-        return self._run_remote_sparse_refit_workers(
-            "stream_remote_sparse_weights",
-            transport=transport,
-            targets=targets,
-            transfer_id=transfer_id,
-            api_key_env_var=api_key_env_var,
-            timeout_s=timeout_s,
-        )
-
-    def finish_remote_sparse_delta_sync(self, succeeded: bool) -> list[ray.ObjectRef]:
-        return self.worker_group.run_all_workers_single_data(
-            "finish_remote_sparse_delta_sync", succeeded=succeeded
-        )
-
-    def _run_remote_sparse_refit_workers(
-        self,
-        method_name: str,
-        **common_kwargs: Any,
-    ) -> list[ray.ObjectRef]:
-        worker_count = len(self.worker_group.workers)
-        return self.worker_group.run_all_workers_multiple_data(
-            method_name,
-            common_kwargs={
-                **common_kwargs,
-                "shard_count": worker_count,
-            },
-            shard_rank=list(range(worker_count)),
-        )
-
     def broadcast_weights_for_collective(
         self, kv_scales: Optional[dict[str, float]] = None
     ) -> list[ray.ObjectRef]:
@@ -1199,7 +1152,7 @@ class Policy(ColocatablePolicyInterface, GenerationInterface):
         the object is lost due to leaving a function scope. It's always recommended that the
         user calls worker_group.shutdown().
         """
-        if not sys.is_finalizing() and hasattr(self, "worker_group"):
+        if hasattr(self, "worker_group"):
             self.worker_group.shutdown(cleanup_method="shutdown")
 
     def start_gpu_profiling(self) -> None:
