+++ pyrefly.toml
@@ -201,6 +201,9 @@ project-includes = [
   "nemo_rl/weight_sync/interfaces.py",
   "nemo_rl/weight_sync/ipc_weight_synchronizer.py",
   "nemo_rl/weight_sync/vllm_remote_sparse_weight_synchronizer.py",
+  "nemo_rl/models/generation/vllm/vllm_sparse_delta.py",
+  "nemo_rl/models/generation/vllm/vllm_sparse_refit.py",
+  "nemo_rl/models/policy/workers/megatron_remote_sparse_refit.py",
   "nemo_rl/utils/weight_transfer_remote_sparse.py",
   "nemo_rl/utils/weight_transfer_sparse_codec.py",
   "nemo_rl/utils/weight_transfer_zmq.py",
