+++ tests/unit/models/generation/test_vllm_generation.py
@@ -12,21 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import asyncio
 import importlib.util
 import json
 import os
 import sys
-import threading
-import time
 import types
-from collections.abc import Iterator
-from concurrent.futures import Future, ThreadPoolExecutor
-from contextlib import contextmanager
 from copy import deepcopy
 from pathlib import Path
-from typing import Any
-from unittest.mock import AsyncMock, MagicMock, call
+from unittest.mock import MagicMock
 
 import pytest
 import ray
@@ -44,8 +37,6 @@ from nemo_rl.models.generation.interfaces import (
 )
 from nemo_rl.models.generation.vllm import VllmConfig, VllmGeneration
 from nemo_rl.models.generation.vllm.vllm_worker import (
-    BaseVllmGenerationWorker,
-    VllmGenerationWorkerImpl,
     _resolve_enable_prefix_caching,
 )
 from nemo_rl.models.generation.vllm.vllm_worker_async import (
@@ -54,18 +45,6 @@ from nemo_rl.models.generation.vllm.vllm_worker_async import (
 )
 from nemo_rl.models.policy import LoRAConfig, PolicyConfig
 from nemo_rl.models.policy.lm_policy import Policy
-from nemo_rl.utils.weight_transfer_remote_sparse import (
-    G_VLLM_REFIT_API_KEY_HEADER,
-    G_VLLM_REFIT_FLUSH_PATH,
-    G_VLLM_REFIT_S3_MANIFEST_PATH,
-)
-from nemo_rl.utils.weight_transfer_zmq import (
-    G_VLLM_REFIT_CHECKSUM_HEADER,
-    G_VLLM_REFIT_PAYLOAD_HEADER,
-    G_VLLM_REFIT_PRODUCER_HEADER,
-    G_VLLM_REFIT_TRANSFER_HEADER,
-    G_VLLM_REFIT_ZMQ_PAYLOAD_PATH,
-)
 
 model_name = "Qwen/Qwen3-0.6B"
 # Define basic vLLM test config
@@ -178,485 +157,6 @@ def test_resolve_enable_prefix_caching_uses_cuda_capability_for_auto(monkeypatch
     monkeypatch.setattr(torch.cuda, "get_device_capability", lambda: (7, 5))
 
     assert _resolve_enable_prefix_caching({}) is False
-
-
-def test_ray_owner_destructors_skip_shutdown_during_interpreter_finalization(
-    monkeypatch,
-):
-    monkeypatch.setattr(sys, "is_finalizing", lambda: True)
-
-    generation = VllmGeneration.__new__(VllmGeneration)
-    generation.shutdown = MagicMock()
-    generation.__del__()
-    generation.shutdown.assert_not_called()
-
-    policy = Policy.__new__(Policy)
-    policy.worker_group = MagicMock()
-    policy.__del__()
-    policy.worker_group.shutdown.assert_not_called()
-
-    cluster = RayVirtualCluster.__new__(RayVirtualCluster)
-    cluster.shutdown = MagicMock()
-    cluster.__del__()
-    cluster.shutdown.assert_not_called()
-
-
-@contextmanager
-def _sparse_refit_worker(
-    *, batch_size: int = 2, futures: list[Future[dict[str, Any]]] | None = None
-) -> Iterator[BaseVllmGenerationWorker]:
-    worker = BaseVllmGenerationWorker.__new__(BaseVllmGenerationWorker)
-    worker._refit_apply_queue_condition = threading.Condition()
-    worker._refit_apply_executor = ThreadPoolExecutor(max_workers=1)
-    worker._refit_apply_futures = list(futures or [])
-    worker._refit_apply_pending_payloads = []
-    worker._refit_seen_payloads = {}
-    worker._refit_apply_queue_depth = 2
-    worker._refit_apply_batch_size = batch_size
-    worker.llm = MagicMock()
-    worker.llm.collective_rpc.return_value = [{"ok": True}]
-    for future in worker._refit_apply_futures:
-        future.add_done_callback(worker._notify_refit_apply_waiters)
-    try:
-        yield worker
-    finally:
-        worker._refit_apply_executor.shutdown(wait=True)
-
-
-def test_sparse_refit_queue_batches_payloads_in_fifo_order() -> None:
-    applied: list[tuple[bytes, ...]] = []
-
-    def apply(payloads: tuple[bytes, ...]) -> dict[str, Any]:
-        applied.append(payloads)
-        return {
-            "ok": True,
-            "payloads": len(payloads),
-            "receiver_total_s": float(len(payloads)),
-        }
-
-    with _sparse_refit_worker(batch_size=3) as worker:
-        worker.update_weights_from_serialized_sparse_payloads = apply
-        responses = [
-            worker._enqueue_sparse_payload_apply(
-                payload, ("transfer", 0, index), str(index)
-            )
-            for index, payload in enumerate((b"0", b"1", b"2", b"3", b"4"))
-        ]
-        response = worker._flush_queued_sparse_payloads()
-        responses.append(response)
-
-    assert applied == [
-        (b"0", b"1", b"2"),
-        (b"3", b"4"),
-    ]
-    assert response["payloads"] == 5
-    assert response["batches"] == 2
-    assert sum(result.get("receiver_total_s", 0.0) for result in responses) == 5.0
-    worker.llm.collective_rpc.assert_called_once_with(
-        "finish_sparse_delta_refit", args=()
-    )
-
-
-def test_sparse_refit_queue_deduplicates_transactional_payloads() -> None:
-    key = ("transfer", 0, 1)
-    with _sparse_refit_worker() as worker:
-        worker.update_weights_from_serialized_sparse_payloads = MagicMock(
-            return_value={"ok": True, "payloads": 1}
-        )
-        assert worker._enqueue_sparse_payload_apply(b"payload", key, "checksum")["ok"]
-        duplicate = worker._enqueue_sparse_payload_apply(b"payload", key, "checksum")
-        assert duplicate == {"ok": True, "payloads": 0, "duplicate": True}
-        with pytest.raises(ValueError, match="reused with different data"):
-            worker._enqueue_sparse_payload_apply(b"other", key, "different")
-        response = worker._flush_queued_sparse_payloads()
-
-    assert response["payloads"] == 1
-    assert worker._refit_seen_payloads == {}
-
-
-def test_sparse_refit_queue_does_not_deduplicate_failed_enqueue() -> None:
-    failed = Future()
-    failed.set_exception(RuntimeError("prior apply failed"))
-    with _sparse_refit_worker(futures=[failed]) as worker:
-        with pytest.raises(RuntimeError, match="prior apply failed"):
-            worker._enqueue_sparse_payload_apply(
-                b"payload", ("transfer", 0, 1), "checksum"
-            )
-
-    assert worker._refit_seen_payloads == {}
-    assert worker._refit_apply_pending_payloads == []
-
-
-def test_sparse_refit_collective_response_merges_verification_metrics() -> None:
-    response = BaseVllmGenerationWorker._refit_collective_response(
-        [
-            {
-                "receiver_total_s": 1.0,
-                "verification_candidates": 4,
-                "verification_samples": 2,
-                "verification_exact_mismatches": 1,
-                "verification_mismatches": 0,
-                "verification_abs_sum": 0.25,
-                "verification_max_abs": 0.25,
-            },
-            {
-                "receiver_total_s": 2.0,
-                "verification_candidates": 4,
-                "verification_samples": 3,
-                "verification_exact_mismatches": 2,
-                "verification_mismatches": 1,
-                "verification_abs_sum": 0.5,
-                "verification_max_abs": 0.4,
-            },
-        ]
-    )
-
-    assert response == {
-        "ok": True,
-        "receiver_total_s": 2.0,
-        "verification_candidates": 4,
-        "verification_samples": 5,
-        "verification_exact_mismatches": 3,
-        "verification_mismatches": 1,
-        "verification_abs_sum": 0.75,
-        "verification_max_abs": 0.4,
-    }
-
-
-def test_sparse_refit_queue_releases_condition_while_backpressured() -> None:
-    first: Future[dict[str, Any]] = Future()
-    second: Future[dict[str, Any]] = Future()
-    started = threading.Event()
-
-    with _sparse_refit_worker(futures=[first, second]) as worker:
-        with ThreadPoolExecutor(max_workers=1) as callers:
-            call = callers.submit(
-                lambda: (
-                    started.set(),
-                    worker._enqueue_sparse_payload_apply(
-                        b"payload", ("transfer", 0, 1), "checksum"
-                    ),
-                )[1]
-            )
-            assert started.wait(timeout=1.0)
-            time.sleep(0.05)
-            assert worker._refit_apply_queue_condition.acquire(timeout=1.0)
-            worker._refit_apply_queue_condition.release()
-            first.set_result({"ok": True, "payloads": 1})
-            assert call.result(timeout=1.0)["ok"]
-
-
-def test_sparse_refit_batch_uses_one_collective_rpc(tmp_path: Path) -> None:
-    worker = VllmGenerationWorkerImpl.__new__(VllmGenerationWorkerImpl)
-    staged_payloads: list[bytes] = []
-
-    def collective_rpc(method, args):
-        assert method == "update_weights_from_sparse_payload_files"
-        staged_payloads.extend(Path(path).read_bytes() for path in args)
-        return [{"ok": True, "receiver_total_s": 1.0}]
-
-    worker.llm = MagicMock(collective_rpc=MagicMock(side_effect=collective_rpc))
-    worker._refit_workers_share_node = True
-    worker._refit_batch_staging_dir = str(tmp_path)
-    payloads = (b"0", b"1", b"2")
-
-    response = worker.update_weights_from_serialized_sparse_payloads(payloads)
-
-    assert staged_payloads == list(payloads)
-    assert not list(tmp_path.iterdir())
-    worker.llm.collective_rpc.assert_called_once()
-    assert response == {"ok": True, "receiver_total_s": 1.0, "payloads": 3}
-
-
-def test_sparse_refit_batch_drains_workers_before_error_cleanup(tmp_path: Path) -> None:
-    worker = VllmGenerationWorkerImpl.__new__(VllmGenerationWorkerImpl)
-    staged_paths: tuple[str, ...] = ()
-
-    def collective_rpc(method, args):
-        nonlocal staged_paths
-        if method == "update_weights_from_sparse_payload_files":
-            staged_paths = args
-            raise RuntimeError("apply failed")
-        assert method == "synchronize_device"
-        assert all(Path(path).is_file() for path in staged_paths)
-        return [True]
-
-    worker.llm = MagicMock(collective_rpc=MagicMock(side_effect=collective_rpc))
-    worker._refit_workers_share_node = True
-    worker._refit_batch_staging_dir = str(tmp_path)
-
-    with pytest.raises(RuntimeError, match="apply failed"):
-        worker.update_weights_from_serialized_sparse_payloads((b"0", b"1"))
-
-    assert [call.args[0] for call in worker.llm.collective_rpc.call_args_list] == [
-        "update_weights_from_sparse_payload_files",
-        "synchronize_device",
-    ]
-    assert not list(tmp_path.iterdir())
-
-
-def test_sparse_refit_batch_falls_back_across_nodes() -> None:
-    worker = VllmGenerationWorkerImpl.__new__(VllmGenerationWorkerImpl)
-    worker._refit_workers_share_node = False
-    worker.llm = MagicMock(
-        collective_rpc=MagicMock(return_value=[{"ok": True, "receiver_total_s": 1.0}])
-    )
-
-    response = worker.update_weights_from_serialized_sparse_payloads((b"0", b"1", b"2"))
-
-    assert worker.llm.collective_rpc.call_args_list == [
-        call("update_weights_from_serialized_sparse_payload", args=(payload,))
-        for payload in (b"0", b"1", b"2")
-    ]
-    assert response == {"ok": True, "receiver_total_s": 3.0, "payloads": 3}
-
-
-@pytest.mark.asyncio
-async def test_async_sparse_refit_batch_bridges_to_async_collective(
-    tmp_path: Path,
-) -> None:
-    worker = VllmAsyncGenerationWorkerImpl.__new__(VllmAsyncGenerationWorkerImpl)
-    staged_payloads: list[bytes] = []
-
-    class AsyncLlm:
-        async def collective_rpc(
-            self, method: str, args: tuple[Any, ...]
-        ) -> list[dict[str, Any]]:
-            assert method == "update_weights_from_sparse_payload_files"
-            staged_payloads.extend(Path(path).read_bytes() for path in args)
-            return [{"ok": True, "receiver_total_s": 1.0}]
-
-    worker.llm = AsyncLlm()
-    worker._refit_async_loop = asyncio.get_running_loop()
-    worker._refit_workers_share_node = True
-    worker._refit_batch_staging_dir = str(tmp_path)
-
-    response = await asyncio.to_thread(
-        worker.update_weights_from_serialized_sparse_payloads,
-        (b"0", b"1"),
-    )
-
-    assert staged_payloads == [b"0", b"1"]
-    assert response == {"ok": True, "receiver_total_s": 1.0, "payloads": 2}
-
-
-@pytest.mark.asyncio
-async def test_sparse_refit_payload_handlers_decode_and_enqueue(monkeypatch) -> None:
-    worker = BaseVllmGenerationWorker.__new__(BaseVllmGenerationWorker)
-    enqueue = MagicMock(
-        side_effect=[
-            {"ok": True, "payloads": 1},
-            {"ok": True, "payloads": 1},
-        ]
-    )
-    worker._enqueue_sparse_payload_apply = enqueue
-    monkeypatch.setattr(
-        "nemo_rl.models.generation.vllm.vllm_worker.download_s3_refit_payload",
-        lambda _manifest: b"s3-payload",
-    )
-    monkeypatch.setattr(
-        "nemo_rl.models.generation.vllm.vllm_worker.decode_sparse_payload",
-        lambda _body, _checksum: b"zmq-payload",
-    )
-
-    s3_result = await worker._apply_s3_manifest_payload(
-        {"key": "object-key", "checksum": "checksum"}
-    )
-    assert s3_result["ok"]
-    assert s3_result["receiver_s3_download_s"] >= 0.0
-
-    invalid_request = types.SimpleNamespace(headers={}, body=AsyncMock())
-    with pytest.raises(ValueError, match="Missing or invalid"):
-        await worker._apply_zmq_payload(invalid_request)
-
-    request = types.SimpleNamespace(
-        headers={
-            G_VLLM_REFIT_TRANSFER_HEADER: "transfer",
-            G_VLLM_REFIT_PRODUCER_HEADER: "2",
-            G_VLLM_REFIT_PAYLOAD_HEADER: "3",
-            G_VLLM_REFIT_CHECKSUM_HEADER: "checksum",
-        },
-        body=AsyncMock(return_value=b"compressed"),
-    )
-    zmq_result = await worker._apply_zmq_payload(request)
-    assert zmq_result["ok"]
-    assert zmq_result["receiver_zmq_decode_s"] >= 0.0
-    assert enqueue.call_args_list == [
-        call(b"s3-payload", ("object-key", -1, -1), "checksum"),
-        call(b"zmq-payload", ("transfer", 2, 3), "checksum"),
-    ]
-
-
-def test_sparse_refit_api_auth_dispatch_and_error_mapping(monkeypatch) -> None:
-    from fastapi import FastAPI
-    from fastapi.testclient import TestClient
-
-    monkeypatch.setenv("NRL_TEST_REFIT_KEY", "secret")
-    worker = BaseVllmGenerationWorker.__new__(BaseVllmGenerationWorker)
-    worker.cfg = {
-        "vllm_cfg": {
-            "async_engine": True,
-            "http_refit_api_key_env_var": "NRL_TEST_REFIT_KEY",
-        }
-    }
-    worker._refit_async_loop = None
-    worker._apply_s3_manifest_payload = AsyncMock(
-        return_value={"ok": True, "payloads": 1}
-    )
-    worker._apply_zmq_payload = AsyncMock(side_effect=RuntimeError("apply failed"))
-    worker._flush_queued_sparse_payloads = MagicMock(
-        return_value={"ok": True, "payloads": 2}
-    )
-    app = FastAPI()
-    worker._setup_vllm_refit_api_server(app)
-    headers = {G_VLLM_REFIT_API_KEY_HEADER: "secret"}
-
-    with TestClient(app) as client:
-        unauthorized = client.post(G_VLLM_REFIT_S3_MANIFEST_PATH, json={})
-        s3_response = client.post(
-            G_VLLM_REFIT_S3_MANIFEST_PATH,
-            json={"key": "key"},
-            headers=headers,
-        )
-        flush_response = client.post(G_VLLM_REFIT_FLUSH_PATH, headers=headers)
-        zmq_response = client.post(
-            G_VLLM_REFIT_ZMQ_PAYLOAD_PATH,
-            content=b"payload",
-            headers=headers,
-        )
-
-    assert unauthorized.status_code == 403
-    assert unauthorized.json() == {"ok": False, "error": "unauthorized"}
-    assert s3_response.status_code == 200
-    assert s3_response.json() == {"ok": True, "payloads": 1}
-    assert flush_response.status_code == 200
-    assert flush_response.json() == {"ok": True, "payloads": 2}
-    assert zmq_response.status_code == 500
-    assert zmq_response.json() == {"ok": False, "error": "apply failed"}
-    assert worker._refit_async_loop is not None
-    worker._apply_s3_manifest_payload.assert_awaited_once_with({"key": "key"})
-    worker._apply_zmq_payload.assert_awaited_once()
-    worker._flush_queued_sparse_payloads.assert_called_once_with()
-
-
-def test_sync_sparse_refit_server_shutdown_cleans_transport_resources(
-    monkeypatch,
-) -> None:
-    import uvicorn
-
-    from nemo_rl.models.generation.vllm import vllm_worker as worker_module
-
-    configs = []
-    servers = []
-
-    def make_config(app, **kwargs):
-        config = types.SimpleNamespace(app=app, **kwargs)
-        configs.append(config)
-        return config
-
-    class Server:
-        def __init__(self, config) -> None:
-            self.config = config
-            self.should_exit = False
-            self.ran = threading.Event()
-            servers.append(self)
-
-        def run(self) -> None:
-            self.ran.set()
-
-    monkeypatch.setattr(uvicorn, "Config", make_config)
-    monkeypatch.setattr(uvicorn, "Server", Server)
-    monkeypatch.setattr(worker_module, "_get_free_port_local", lambda *_args: 12345)
-    monkeypatch.setattr(worker_module, "_get_node_ip_local", lambda: "10.0.0.1")
-    collect = MagicMock()
-    empty_cache = MagicMock()
-    monkeypatch.setattr(worker_module.gc, "collect", collect)
-    monkeypatch.setattr(worker_module.torch.cuda, "empty_cache", empty_cache)
-
-    worker = VllmGenerationWorkerImpl.__new__(VllmGenerationWorkerImpl)
-    worker.cfg = {
-        "vllm_cfg": {"http_refit_server_port": None},
-        "port_range_low": 10000,
-        "port_range_high": 11000,
-    }
-    worker._setup_vllm_refit_api_server = MagicMock()
-    worker._refit_http_server = None
-    worker._setup_vllm_refit_server()
-
-    assert len(configs) == 1
-    assert configs[0].host == "0.0.0.0"
-    assert configs[0].port == 12345
-    assert servers[0].ran.wait(timeout=1.0)
-    assert worker.report_refit_server_base_url() == "http://10.0.0.1:12345"
-    worker._setup_vllm_refit_api_server.assert_called_once_with(configs[0].app)
-
-    relay = MagicMock()
-    llm = MagicMock()
-    worker._zmq_refit_server = (relay, "tcp://relay")
-    worker._flush_queued_sparse_payloads = MagicMock()
-    worker._refit_apply_executor = MagicMock()
-    worker.llm = llm
-    worker.tokenizer = object()
-
-    assert worker.shutdown() is True
-    relay.close.assert_called_once_with()
-    worker._flush_queued_sparse_payloads.assert_called_once_with()
-    worker._refit_apply_executor.shutdown.assert_called_once_with(wait=True)
-    assert servers[0].should_exit is True
-    assert worker._refit_http_server is None
-    llm.collective_rpc.assert_called_once_with("cleanup", args=())
-    assert worker.llm is None
-    assert worker.tokenizer is None
-    collect.assert_called_once_with()
-    empty_cache.assert_called_once_with()
-
-
-@pytest.mark.asyncio
-async def test_async_sparse_refit_post_init_records_worker_locality() -> None:
-    worker = VllmAsyncGenerationWorkerImpl.__new__(VllmAsyncGenerationWorkerImpl)
-    worker.cfg = {"refit_transport": "vllm_zmq_sparse"}
-    worker._mtp_load_from_disk = False
-    worker.report_device_id_async = AsyncMock(return_value=["0"])
-    worker.llm = MagicMock()
-    worker.llm.collective_rpc = AsyncMock(return_value=["node-0", "node-0"])
-
-    await worker.post_init_async()
-
-    assert worker.vllm_device_ids == ["0"]
-    assert worker._refit_workers_share_node is True
-    assert worker.llm.collective_rpc.await_args_list == [
-        call("bind_numa", args=()),
-        call("report_node_hostname", args=()),
-    ]
-
-
-def test_async_sparse_refit_exposes_zmq_relay(monkeypatch) -> None:
-    server = MagicMock()
-    server_type = MagicMock(return_value=server)
-    monkeypatch.setattr(
-        "nemo_rl.models.generation.vllm.vllm_worker.ZmqSparseRefitServer",
-        server_type,
-    )
-    monkeypatch.setattr(
-        "nemo_rl.models.generation.vllm.vllm_worker._get_free_port_local",
-        lambda *_args: 12345,
-    )
-    monkeypatch.setattr(
-        "nemo_rl.models.generation.vllm.vllm_worker._get_node_ip_local",
-        lambda: "10.0.0.1",
-    )
-    worker = VllmAsyncGenerationWorkerImpl.__new__(VllmAsyncGenerationWorkerImpl)
-    worker.cfg = {"vllm_cfg": {"zmq_refit_server_port": None}}
-    worker._zmq_refit_server = None
-
-    assert worker.start_zmq_sparse_refit_relay(["http://receiver"]) == (
-        "tcp://10.0.0.1:12345"
-    )
-    server.start.assert_called_once_with()
-
-    worker.stop_zmq_sparse_refit_relay()
-    server.close.assert_called_once_with()
-    assert worker._zmq_refit_server is None
 
 
 basic_lora_test_config: LoRAConfig = {
